Major refactor and update to middleware, MultiAuth and HawkAuth backends
Breaking Changes
==================

Changes to authenticate() return value
-------------------------------------

This updates the backend interface to allow backends to return more than just the `user` object.
Some backends, like the Hawk backend, do more than just authentication. They can extract extra
information from the `Authorization` header that may be of importance to the request.

So now, the result of a successful authentication will be provided to the `req` object as follows:

```
	# Hawk auth example
	req.context['auth'] = {
		# Always present
		'backend': <the backend instance that authenticated the user>,
		'user': <user object returned from user_loader()>,

		# Backend-specific, if provided by backend
		'receiver': <the Hawk Receiver object used to authenticate the request>,
	}
```

Notice that the `'user'` key is nested one level deeper under the `'auth'` dictionary. A user of
this library can now specify the key to hold the authentication results in the `req.context`
dictionary. The default value is `'auth'`, as seen above.

Anyone writing their own backends will need to modify their backend's `authenticate()` method to
return a dictionary like the above format which includes at least the `'user'` key.

Changes to authenticate() semantics
-----------------------------------

Custom `authenthicate()` implementations should always either return a dictionary containing the
user object, or raise a `UserNotFound` exception. One can use the new `AuthBackend.load_user()`
method to accomplish this. It invokes the provided `user_loader` callback and raises a
`UserNotFound` exception if the user cannot be retrieved from the app. In this way, we can simplify
any `authenticate()` implementation by writing code that assumes the user has been found, as the
`UserNotFound` exception will be caught by the middleware.

Changes to HawkAuthBackend interface
--------------------------------------

The `HawkAuthBackend` now takes a second parameter `credentials_loader` which will receive the
retrieved user object and is expected to return the Hawk credentials map data. It is now an error to
provide the `credentials_map` value in the `receiver_kwargs` dictionary.

Changes to MultiAuthBackend
-----------------------------

There is a new parameter to the `MultiAuthBackend` constructor: `early_exit`. If true, the
`MultiAuthBackend` will stop iterating through its list of backends and exit with a failure to
authenticate if the current backend raises a non-`BackendNotApplicable`
`BackendAuthenticationFailure` exception. This is to allow a backend to "claim" the authentication
of a request and assert that it was able to examine the request and definitively say that the
credentials were invalid for the indicated user. Exceptions that will ignore the `early_exit`
setting and allow the `MultiAuthBackend` to continue iterating are:

- `UserNotFound`
- `BackendNotApplicable`
- Any other `falcon.HTTPUnauthorized` exception
- Any other `Exception` (though this will likely result in an end to iteration anyway as the
  exception bubbles up through the stack)

In situations where the `MultiAuthBackend` is allowed to iterate through the entire list of backends
without success, any challenges presented by those backends (in the form of the `WWW-Authenticate`
header on the `falcon.HTTPUnauthorized` exceptions) will still be collected and included in the
ultimate `falcon.HTTPUnauthorized` exception raised by the `MultiAuthBackend`.

Changes to error messages
-------------------------

Many error titles and descriptions have been updated for consistency and formatting. For instance,
"Invalid authentication backend" errors to no longer include backticks in the string.

Benign Changes
==============

Success/failure callbacks
-------------------------

Added support for `on_success()` and `on_failure()` callbacks

This allows one to provide callbacks to the middleware that will be invoked upon success or failure
to authenticate. This is to allow the application to have more insight into the inner workings of
the auth process (read: logging).

The `on_success` callback will receive the `req`, `resp`, `resource` and `results` parameters.
The `on_failure` callback will receive the `req`, `resp`, `resource` and `exception` parameters. The
`exception` parameter will be of type `BackendAuthenticationFailure` which contains information
about the backend that failed the authentication.