**Breaking Changes**
Update the backend interface to allow backends to return more than just the `user` object. Some
backends, like the Hawk backend, do more than just authentication. They can extract extra
information from the `Authorization` header that may be of importance to the request.

So now, the result of a successful authentication will be provided to the `req` object as follows:

```
	# Hawk auth example
	req.context['auth'] = {
		'type': 'hawk',
		'user': <user object returned from user_loader()>,
		'receiver': <the Hawk Receiver object used to authenticate the request>,
	}
```

Anyone writing their own backends will need to make the following changes:

- Add `AUTH_TYPE` as a class member on their AuthBackend subclass
- Return a dictionary like the above format which includes at least the `'user'` key

These changes also allow a user of this library to specify the key to use in the `req.context`
dictionary. The default value is `'auth'`, as seen above.

Finally, this also updates the "Invalid authentication backend" errors to no longer include
backticks in the string.