Add support for on_success() and on_failure() callbacks
This allows one to provide callbacks to the middleware that will be invoked upon success or failure
to authenticate. This is to allow the application to have more insight into the inner workings of
the auth process (read: logging).

The `on_failure` callback will only be called when the backend was appropriate for
the authentication credentials on the request and they were invalid for some reason. For example, if
you're using a Hawk auth backend and a request comes in with Basic auth credentials, the
`on_failure` callback will not be called. This becomes more relevant when using the
`MultiAuthBackend`, where you wouldn't want to have multiple `on_failure` calls as it works through
the list of supported backends.