Use req.uri_template when disabling a resource.
The logic used to disable the auth middleware for a resource does so by adding
the current request's path to the `exempt_routes` collection. This is a somewhat
indirect way of informing the middleware that the resource should not be
processed.

Incidentally, we introduced a bug in a recent commit when we began
using the `req.uri_template` rather than the rendered `req.path` when matching
`exempt_routes`. Unfortunately, we missed the usage in `_get_auth_settings()`
which contains the logic mentioned above.

This PR attempts to address this shortcoming and adds a test to cover this gap.